Moved the experimental usb scan to a new experimental command 'api labs'
